Issue #6556 Ensure context classloader set when operating on memcache.
