+++ jetty-memcached/jetty-memcached-sessions/src/main/java/org/eclipse/jetty/memcached/session/MemcachedSessionDataMap.java
@@ -21,6 +21,7 @@ import java.util.List;
 import net.rubyeye.xmemcached.MemcachedClient;
 import net.rubyeye.xmemcached.XMemcachedClientBuilder;
 import net.rubyeye.xmemcached.transcoders.SerializingTranscoder;
+import org.eclipse.jetty.server.session.RunnableResult;
 import org.eclipse.jetty.server.session.SessionContext;
 import org.eclipse.jetty.server.session.SessionData;
 import org.eclipse.jetty.server.session.SessionDataMap;
@@ -43,6 +44,7 @@ public class MemcachedSessionDataMap extends AbstractLifeCycle implements Sessio
     protected int _expirySec = 0;
     protected boolean _heartbeats = true;
     protected XMemcachedClientBuilder _builder;
+    protected SessionContext _context;
 
     /**
      * SessionDataTranscoder
@@ -140,9 +142,13 @@ public class MemcachedSessionDataMap extends AbstractLifeCycle implements Sessio
     @Override
     public void initialize(SessionContext context)
     {
+        if (isStarted())
+            throw new IllegalStateException("Context set after MemcachedSessionDataMap started");
+        
         try
         {
-            _builder.setTranscoder(new SessionDataTranscoder());
+            _context = context;
+            _builder.setTranscoder(new SessionDataTranscoder(_context.getContext().getClassLoader()));
             _client = _builder.build();
             _client.setEnableHeartBeat(isHeartbeats());
         }
@@ -155,14 +161,47 @@ public class MemcachedSessionDataMap extends AbstractLifeCycle implements Sessio
     @Override
     public SessionData load(String id) throws Exception
     {
-        SessionData data = _client.get(id);
-        return data;
+        if (!isStarted())
+            throw new IllegalStateException("Not started");
+        
+        final RunnableResult<SessionData> result = new RunnableResult<>();
+
+        Runnable r = () ->
+        {
+            try
+            {
+                result.setResult(_client.get(id));
+            }
+            catch (Exception e)
+            {
+                result.setException(e);
+            }
+        };
+
+        _context.run(r);
+        return result.getOrThrow();
     }
 
     @Override
     public void store(String id, SessionData data) throws Exception
     {
-        _client.set(id, _expirySec, data);
+        if (!isStarted())
+            throw new IllegalStateException("Not started");
+        
+        final RunnableResult<Object> result = new RunnableResult<>();
+        Runnable r = () ->
+        {
+            try
+            {
+                _client.set(id, _expirySec, data);
+            }
+            catch (Exception e)
+            {
+                result.setException(e);
+            }
+        };
+        _context.run(r);
+        result.throwIfException();
     }
 
     @Override
